refactor: convert money fields to integer cents (BIGINT) across database, backend, and web frontend

+++ migrations/004_create_menu_items.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS menu_items (
     category_id UUID NOT NULL REFERENCES categories(id) ON DELETE CASCADE,
     name VARCHAR(200) NOT NULL,
     description TEXT,
-    price DECIMAL(10,2) NOT NULL,
+    price BIGINT NOT NULL,
     image_path TEXT,
     is_available BOOLEAN NOT NULL DEFAULT true,
     sort_order INT NOT NULL DEFAULT 0,
